Fixed DRAM write timing.
Not the best fix (as it mocks with nCAS instead of nWE) but works,
it seems. I'm also tired of trying to find a better, cleaner solution.

This is after all not all that relevant: it's just an emulated DRAM
device, so anything that works well enough is good enough.