Much improved and cleaned up ROM.
Now, the default ROM code jumps to TASK mode in DRAM.
After that returns to ROM, it dumps all the registers as well
as ecause eaddr and tpc. This is a great tool to find issues.

There's also a version of ROM (rom_c) that executes C code in
ROM instead of DRAM.

NOTE: all this works, because our simulated ROM is actually SRAM,
thus we can place R/W data in there.

Also checked in is a new app, (memtest) that does some primitive
DRAM confidence tests.