Added DRAM constant, changed DRAM link location.
This later change is important because we're in TASK mode now.
That means that mem_limit registers are potentially set up to be
at the top of DRAM, which must be beyond the peripherals.

Now, that's normally the case, except for the wait-state aliases.

Luckily the wait-state setting is ignored for DRAM, so it doesn't
matter which alias we're actually using for DRAM.