Fixed eaddr issue with back-to-back exceptions.
Let's imagine to back-to-back instructions, each raising
an exception. The first instruction should raise, and the
second one should gen cancelled. All that was handled
right, except that the eaddr (effective address really) of
the second instruction snuck through and got registered
as the eaddr CSR. This fix blocks the capturing of the
values of the second - cancelled - instruction in the
's1' stage of execute.