Removed WS alias from bounds checks.
memory base and limit registers are now leaving the top 4 bits
of the logical address alone when translating to physical address
and checking for AVs. This allows for code to have access to a
piece of the address space in any WS alias.

There also was (I think) a bug in the way base and limit calculations
were performed for fetches: $pc and $tpc are in 16-bit quantities.
This is patched up as well.