New SW infrastructure.
This is the beginnings of a true executive.

We have a well (better) partitioned ROM source code base.
We detect RAM size (albeit single-bank so far), set up a protected
area for scheduler mode storage at the top of DRAM and then
jump to DRAM in task mode. We handle (some) syscalls, even
initialize the console to 115200 baud communication.

We seem to be able to execute simple C code in 'exclusive' mode,
that is TASK mode with full access to all resources.