Reversed default power-on.
The problem with defaulting to power-on-by-default is that there is a short
on transient while the MCU is booting and gets to override the pull-up. This
is mitigated now by a pull-down and an override jumper. The existence of the
jumper now necessitates an open-emitter drive from the MCU; that is, drive-high
for ON and float for OFF.