Added testing against reference (old) implementation
RF is the first test subject in this effort:
the new modules should be drop-in replacements
for the old one, unless something major happens.
As such, it should be possible to create automated
unit-tests that compare the old and the new and complain.

In this checking the test performed is still rather limited
and the framework is local, but it's a step in that direction.