WIP - Started working on new decode
This version has an address generator
for the decode ROM (which should be 1KWord large) but not the content generator for it.

Looking at the compression logic, not sure it's going to be any better than the direct-decode version, especially since this monster will have to sit in the 'logic' domain of the fetch unit.