More branch-related fixes.
Now we have issues with execution in the branch-shadow.

WE are properly cancelling the instruction right
after a branch, but we also are cancelling the one
after that. This is not what the TB assumes.

However, it might be what we want as the behavior: when do_branch goes high, every stage
should drop what they are doing, including exec1.
And since do_branch is registered, that would mean
dropping the instruction after-after the branching
one.