Quite a bit of progress towards self-checking bus_if_tb.
This one self-checks FETCH/MEM/DMA operations.
No check yet for refresh cycles, also no check that
reads return the proper expected values.

There is also an ugly hack where the TB waits an epsilon
to ensure that addr is stable on the falling edge of CAS.
I'm aware of this race, in real HW, the CAS signals are
re-registered to ensure proper sequencing, but this should
not happen.