Things are getting ugly and hacky again.
So, it turns out, I totally don't understand how to break
bursts. I think I'm trying to be a bit too aggressive on latency,
but maybe that's not the main issue.

It also happens that I'm eating bursts (see 4th transaction which has a second burst). I'm also not holding CAS after last beat in this case, but since it should't be the last beat to begin
with, maybe that's a problem for later.